requirements.txt
fastapi
python-telegram-bot>=20.0
httpx
uvicorn
redis